feat: add `IntoRow` and `Schema` derive macros
